refactor(model): split `post.go` into multiple files
